Fix handling of extra schema fields
To avoid ambiguities and inconsistencies in how "extra" fields on metadata record package objects are updated, we mandate - via schema validators - that they must not be missing from the input dictionary. This involves a bit of hackery in the case of list fields (in particular, infrastructures on metadata_record), because CKAN (a) doesn't inherently support validating the presence/absence of lists, and (b) treats empty lists weirdly.