Make schema_version optional
For compatibility with the Plone back-end, which requires only that schema name be supplied when submitting metadata.

Note that the column schema_version on table metadata_schema is left defined as 'not null'. This is fine because a 'null' schema version is actually written as an empty string, which furthermore avoids unique constraint confusion with nulls vs empty strings.