Persist a metadata record's validation state explicitly
We now store the validation state explicitly on a metadata record (as a package extra), instead of trying dynamically to determine whether validation has been done or is required based on complicated interrogation of past validation activities and checking of metadata model and metadata record revision ids. This requires us to explicitly deal with cases where metadata records must be invalidated (i.e. set validation state to 'not validated'); for example, when changing the organization linked to a metadata model, we invalidate all records that are no longer dependent, or that are newly dependent, on that model. This is a far better approach by virtue of both its simplicity and transparency.