Eliminate unnecessary invalidation of metadata records
Previously we invalidated a metadata record any time the owner_org or infrastructure list changed; however, such changes won't necessarily result in a different set of validation models for that record. Instead we get the validation model list before and after updating the record, and invalidate the record only if that list has changed.