Implement organization_delete API override; fixes #10
The intention originally was to chain this action to CKAN's organization_delete, i.e. only implement metadata framework-related dependency checks and cascaded deletes, and then allow CKAN to perform the actual organization delete and anything else that goes with it. However, action chaining only works when the original action (or a CKAN override) has been defined by some other plugin. So instead we do the delete of organization and related stuff directly.