Implement metadata JSON validation
Quite a bit of refactoring here too. We now store a 'validated' flag and 'errors' dictionary on the metadata record instead of 'validation_state', since there is no way to unambiguously decide whether something is invalid or partially valid (at least not without stepping on the toes of the workflows subsystem). The implication is that the metadata record and validation subsystem have no inherent understanding of how valid a metadata record is - they just tell us whether it's been validated or not and what the errors were as per the associated models. The analysis of validity is deferred to the workflows subsystem.

Also revised what and how we log to the activity stream.