Extensive re-work of the workflows subsystem
The main thrust of this change is that metrics affecting workflow transitions must be pushed to the system - as workflow annotations - and validated passively when performing a workflow transition, rather than actively calculating metrics at transition time. We leverage JSON Schema for checking that a metadata record - augmented with annotations - passes the rules for entering a workflow state. In case we do need to calculate any metrics on the fly at transition time, we can extend the JSON Schema for workflow rules with keywords and corresponding processors for calling external services or such.