Improve logging of metadata validation and workflow activities
Removed explicit logging of object revision ids, as this is useless in some cases - e.g. package and package_extra revision ids may differ, meaning you cannot necessarily go directly to the state of say metadata json given a package revision id. The timestamp automatically logged with an activity record can instead be used to determine what the state of an object was at the time of an activity.