Put object names rather than ids in output dicts
for nested objects. This makes for easier testing and debugging and might even be useful in the end-user UI too. Closes #32, although it does not address the potential problem points. However, 1. the increased load time is likely to be negligible, and in any case the relevant objects are likely already cached in memory by SQLAlchemy; and 2. consideration of caching of output dicts by hypothetical external systems can be dealt with in future if the need arises.