Send DOI explicitly as 'record_id' to search agent
Actually, we're sending the metadata_record.name, but normally this would have been set to the DOI (during validation, if the 'published' workflow state has been setup so) by the time the record gets sent for indexing.

We're doing this (rather than just letting the search agent pick the identifier out of the metadata JSON) because we can control the record name (e.g. prevent it being updated once the record has been published) whereas we cannot control changes to a deeply nested field in the metadata - and therefore run the risk of not being able to update or delete the correct record in the index.