Reinstate metadata validity check action
This side effect-free action was previously removed due to the introduction of validation tasks which can cause updates to DB objects. We reintroduce it, however, as it can be useful to a client, and ensure that validation tasks may only run if 'allow_side_effects' is set in the context - which we do within metadata_record_validate, but not in (the still side effect-free) metadata_validity_check.