Don't consider deleted names to be in use
Ref #62

At the model save level we already use the undelete-and-update technique when attempting to create an object with the same name (or same set of values for a unique key, generally) as a previously deleted one. The problem was in the schema validator code.

Note: the problem still exists for CKAN group-type objects (metadata collections etc) because we are still using native CKAN validators for checking objects names.