Add a metadata schema for validating historical SADCO metadata
Identical to the SANS 1878 (MIMS) schema but with a few properties made non-mandatory. In order for the complicated SANS 1878 to DataCite mappings to be written to a metadata record, validation has to complete successfully. It's insufficient to allow certain validation errors through when publishing, using a custom workflow state - because we'll be trying to publish a record without the DataCite properties. Hence we also remove the workflow state added in the previous commit.