Implement role-based access control for all create/update/delete APIs
We take a much simpler approach than the no-longer-used roles plugin in the ckanext-accesscontrol extension. User privileges are saved to Redis by ckanext-accesscontrol upon receipt of an access token, as well as when a user logs in to the CKAN UI. The privileges required for an API are parameterized by the relevant API auth function, and these are checked against the user's privileges which we load from Redis as needed.