Add more detail to immutableResource and linkedResources
Application users want to be able to filter and search by file type, including file types within an archive. This is now supported by immutableResource, in which resourceURL has been replaced with a more detailed resourceDownload object.

Also, added a serverType to linkedResources, which must be filled in if the linkedResourceType is "Query". This enables applications to know what kind of data they're going to get back for a query. We can extend the serverType vocabulary as needed.