Commit validation effects even if validation fails
This has a profound implication for schemas that perform mappings (e.g. from SANS 1878 to SAEON-DataCite):

By allowing the mappings to be committed regardless of validation errors, we can go on to publish, say, historical metadata that does not fully comply with the schema, by allowing specific errors through an appropriate workflow state, instead of having to create a near duplicate of a highly complex schema simply to relax some of the mandatory field requirements.

We also make sure that empty fields are properly discarded from the input metadata.