Fix race condition
The "mapInit" keyword must be processed before any "mapTo" keywords. The order of keywords in the schema depends unpredictably on keyword names, and it just so happens that "mapInit" has always appeared before "properties" at the top level. But this is not guaranteed, and experimenting with different keyword names I got "properties" to appear before the initialization keyword, which broke the schema mapping process.